+++ src/schema.rs
@@ -167,6 +167,7 @@ pub(crate) fn example_road_edge() -> RoadEdge<f64> {
         )),
         Flow(0.4),
         Time(4.0),
+        true,
     )
 }