+++ tests/legs.rs
@@ -42,6 +42,7 @@ fn get_simulation() -> Simulation<f64> {
                 SpeedDensityFunction::FreeFlow,
                 Flow::infinity(),
                 Time(0.),
+                true,
             ),
         ),
         (
@@ -54,6 +55,7 @@ fn get_simulation() -> Simulation<f64> {
                 SpeedDensityFunction::FreeFlow,
                 Flow::infinity(),
                 Time(0.),
+                true,
             ),
         ),
     ]);