+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The tag [OUTPUT] indicates changes affecting the output files.
 
 - [USER] Add road-network parameter `max_pending_duration` to specify the delay after which vehicles
   are forced to enter the next edge in case of spillback.
+- [USER] Add edge parameter `overtaking` to allow vehicles to overtake each other at the edge's exit
+  bottleneck.
 - [USER] Add the current time of the events in the within-day model progress bar.
 - [USER] Ignore warnings when more than 20 have already been emitted.
 - [DEV] Add a debug check that all agents have reached their destination.